+++ NEWS.md
@@ -1,6 +1,7 @@
 # Version 5.2.2
 
 - Skip more tests on CRAN. White-list tests instead of blacklisting them in order to try to keep check time under the official 10-minute cap.
+- Disallow wildcard names to grep-match other wildcard names or any replacement values. This will prevent careless mistakes and confusion when generating `drake_plan()`s.
 
 # Version 5.2.1